+++ include/soralog/impl/fallback_configurator.hpp
@@ -7,10 +7,10 @@
 
 #pragma once
 
-#include <tuple>
 #include <functional>
 #include <soralog/configurator.hpp>
 #include <soralog/logging_system.hpp>
+#include <tuple>
 
 namespace soralog {
 
@@ -38,7 +38,7 @@ namespace soralog {
 
     ~FallbackConfigurator() override = default;
 
-    void prepare(LoggingSystem &system, int index, Result &result) override;
+    void prepare(LoggingSystem &system, size_t index, Result &result) override;
     void applySinks() const override;
     void applyGroups() const override;
     void cleanup() override;
@@ -46,7 +46,10 @@ namespace soralog {
    private:
     Level level_ = Level::INFO;  ///< Default logging level.
     bool with_color_ = false;    ///< Enables colored console output.
-    std::optional<std::tuple<std::reference_wrapper<LoggingSystem>, int, std::reference_wrapper<Result>>> applicator_;
+    std::optional<std::tuple<std::reference_wrapper<LoggingSystem>,
+                             size_t,
+                             std::reference_wrapper<Result>>>
+        applicator_;
   };
 
 }  // namespace soralog
